chore: Add base_image configuration option to dagger
This commit adds a new configuration option `base_image` to the `dagger` application. The `base_image` option allows users to specify the base image for the Docker container used by the application.

The code changes include:
- Adding the `base_image` field to the `Configuration` struct in `configuration.rs`
- Modifying the `main` function in `main.rs` to pass the `base_image` value to the `echo` function in the `utils` module
- Adding a new `utils` module in `steps.rs` to contain the `echo` function

This change improves the flexibility of the `dagger` application by allowing users to easily customize the base image used for the Docker container.